+++ app/models/performed_segment.rb
@@ -3,4 +3,6 @@ class PerformedSegment < ApplicationRecord
   validates  :starting_time, presence: true
 
   belongs_to :competition
+  belongs_to :category
+  belongs_to :segment
 end