+++ spec/controllers/elements_controller_spec.rb
@@ -49,8 +49,7 @@ RSpec.describe ElementsController, type: :controller do
       end
     end
     describe 'sort:' do
-      datatable.column_names.each do |key|
-      #[:level].each do |key|
+      datatable.columns.select(&:orderable).map(&:name).each do |key|
         it key do
           expect_order(elem4T, elemLSp, key)
         end