+++ spec/updaters/skaters_updater_spec.rb
@@ -8,9 +8,9 @@ RSpec.describe Skater, updater: true do
   before(:all) {
     SkaterUpdater.new.update_skaters
   }
-  [:MEN, :LADIES, :PAIRS, :"ICE DANCE"].each do |category|
-    context "\# of skater in '#{category}'" do
-      it { expect(Skater.where(category: category).count).to be > 0 }
+  [:MEN, :LADIES, :PAIRS, :"ICE DANCE"].each do |category_str|
+    context "\# of skater in '#{category_str}'" do
+      it { expect(Skater.where(category: Category.find_by(name: category_str)).count).to be > 0 }
     end
   end
 end